refactor: move server into a directory
